Refactor ImplementationSnapshot hierarchy
