+++ subprojects/execution/src/main/java/org/gradle/internal/execution/history/changes/DefaultExecutionStateChangeDetector.java
@@ -24,7 +24,7 @@ import org.gradle.internal.Cast;
 import org.gradle.internal.execution.history.BeforeExecutionState;
 import org.gradle.internal.execution.history.PreviousExecutionState;
 import org.gradle.internal.snapshot.FileSystemSnapshot;
-import org.gradle.internal.snapshot.impl.KnownImplementationSnapshot;
+import org.gradle.internal.snapshot.impl.ClassImplementationSnapshot;
 
 import static org.gradle.internal.execution.history.impl.OutputSnapshotUtil.findOutputsStillPresentSincePreviousExecution;
 
@@ -45,8 +45,8 @@ public class DefaultExecutionStateChangeDetector implements ExecutionStateChange
         // After validation, the current implementations can't be unknown when detecting changes.
         // Previous implementations can still be unknown, since we store the inputs in the task history even if validation fails.
         // When we fail the build for unknown implementations, then the previous implementations also can't be unknown.
-        KnownImplementationSnapshot currentImplementation = Cast.uncheckedNonnullCast(thisExecution.getImplementation());
-        ImmutableList<KnownImplementationSnapshot> currentAdditionalImplementations = Cast.uncheckedNonnullCast(thisExecution.getAdditionalImplementations());
+        ClassImplementationSnapshot currentImplementation = Cast.uncheckedNonnullCast(thisExecution.getImplementation());
+        ImmutableList<ClassImplementationSnapshot> currentAdditionalImplementations = Cast.uncheckedNonnullCast(thisExecution.getAdditionalImplementations());
         ChangeContainer implementationChanges = new ImplementationChanges(
             lastExecution.getImplementation(), lastExecution.getAdditionalImplementations(),
             currentImplementation, currentAdditionalImplementations,
