+++ subprojects/snapshots/src/main/java/org/gradle/internal/snapshot/impl/LambdaImplementationSnapshot.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 2020 the original author or authors.
+ * Copyright 2022 the original author or authors.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -20,60 +20,111 @@ import org.gradle.internal.hash.HashCode;
 import org.gradle.internal.hash.Hasher;
 
 import javax.annotation.Nullable;
+import java.lang.invoke.SerializedLambda;
+import java.util.Objects;
 
 public class LambdaImplementationSnapshot extends ImplementationSnapshot {
 
-    public LambdaImplementationSnapshot(String typeName) {
-        super(typeName);
+    private final HashCode classLoaderHash;
+    private final String implClass;
+    private final String implMethodName;
+    private final String implMethodSignature;
+    private final int implMethodKind;
+
+    public LambdaImplementationSnapshot(HashCode classLoaderHash, SerializedLambda lambda) {
+        this(
+            lambda.getCapturingClass(),
+            classLoaderHash,
+            lambda.getImplClass(),
+            lambda.getImplMethodName(),
+            lambda.getImplMethodSignature(),
+            lambda.getImplMethodKind()
+        );
+    }
+
+    public LambdaImplementationSnapshot(
+        String capturingClass,
+        HashCode classLoaderHash,
+        String implClass,
+        String implMethodName,
+        String implMethodSignature,
+        int implMethodKind
+    ) {
+        super(capturingClass);
+        this.classLoaderHash = classLoaderHash;
+        this.implClass = implClass;
+        this.implMethodName = implMethodName;
+        this.implMethodSignature = implMethodSignature;
+        this.implMethodKind = implMethodKind;
     }
 
     @Override
     public void appendToHasher(Hasher hasher) {
-        throw new RuntimeException("Cannot hash implementation of lambda " + getTypeName());
+        hasher.putString(ImplementationSnapshot.class.getName());
+        hasher.putString(getTypeName());
+        hasher.putHash(classLoaderHash);
+        hasher.putString(implClass);
+        hasher.putString(implMethodName);
+        hasher.putString(implMethodSignature);
+        hasher.putInt(implMethodKind);
+    }
+
+    @Nullable
+    @Override
+    public HashCode getClassLoaderHash() {
+        return classLoaderHash;
+    }
+
+    public String getImplClass() {
+        return implClass;
+    }
+
+    public String getImplMethodName() {
+        return implMethodName;
+    }
+
+    public String getImplMethodSignature() {
+        return implMethodSignature;
+    }
+
+    public int getImplMethodKind() {
+        return implMethodKind;
+    }
+
+    @Override
+    public boolean isUnknown() {
+        return false;
+    }
+
+    @Nullable
+    @Override
+    public UnknownReason getUnknownReason() {
+        return null;
     }
 
     @Override
     protected boolean isSameSnapshot(@Nullable Object o) {
+        return equals(o);
+    }
+
+    @Override
+    public boolean equals(Object o) {
         if (this == o) {
             return true;
         }
         if (o == null || getClass() != o.getClass()) {
             return false;
         }
-
         LambdaImplementationSnapshot that = (LambdaImplementationSnapshot) o;
-
-        return getTypeName().equals(that.getTypeName());
-    }
-
-    @Override
-    public HashCode getClassLoaderHash() {
-        return null;
-    }
-
-    @Override
-    public boolean isUnknown() {
-        return true;
-    }
-
-    @Override
-    @Nullable
-    public UnknownReason getUnknownReason() {
-        return UnknownReason.LAMBDA;
-    }
-
-    @Override
-    public boolean equals(Object o) {
-        return false;
+        return implMethodKind == that.implMethodKind &&
+            classLoaderHash.equals(that.classLoaderHash) &&
+            implClass.equals(that.implClass) &&
+            implMethodName.equals(that.implMethodName) &&
+            implMethodSignature.equals(that.implMethodSignature);
     }
 
     @Override
     public int hashCode() {
-        return getTypeName().hashCode();
-    }
-
-    @Override
-    public String toString() {
-        return getTypeName();
+        return Objects.hash(classLoaderHash, implClass, implMethodName, implMethodSignature, implMethodKind);
     }
 }
